Set up Unity gitignore file and clear out files
I just set it with the Unity defaults for now, which I think covers the
set of generated and transient files in the tree. However, it may be
worth considering keeping the project files around in the future, as
they can be used outside of the context of Unity.

For reference, the file purge was completed with:
  git rm --cached -r .
  git add .