Changed NES tile handling to use XML serialization
The NES tile bytes are now mapped via XML serialization. Changes made using the new Prefab Mapper are instantly rendered in the mapper as well as saved simulatanously to the /assets/resources/nesTilesMap.xml configuration file. 

After changes are made to the nesTilesMap.xml you will want to regenerate the level. 

Individual changes can be made with the other menu option "Prefab Replacement" when that tool is used you can simply save the scene after the changes are applied.