Added in duplicate tile removal
Duplicate tile removal is working with about 90% effectiveness. Some tiles are still left as duplicates, but I think it is in cases where there were 3 layers of duplicates or more instead of 2. Should be a lot less work to do now. 

When you regenerate all of brinstar with this it will take roughly 4-5 minutes. All the speed gains we got on the last update were spent on duplicate tile removal searching. :)